ajustes

+++ databases/database-load/UseCases/ClearDatabaseUseCase.cs
@@ -5,13 +5,24 @@ namespace database_load_playground.UseCases;
 
 public static class ClearDatabaseUseCase
 {
-    public static async Task ExecuteAsync()
+    public static async Task ExecuteAsync(bool clearTransacao, bool clearTransacaoPart)
     {
         AnsiConsole.Markup("[gray]Limpar banco de dados...[/]");
         await using var conn = await DbFactory.CreateConnectionAsync();
         await using var cmd = conn.CreateCommand();
-        cmd.CommandText = "truncate transacao";
-        await cmd.ExecuteNonQueryAsync();
+
+        if (clearTransacao)
+        {
+            cmd.CommandText = "truncate transacao";
+            await cmd.ExecuteNonQueryAsync();
+        }
+
+        if (clearTransacaoPart)
+        {
+            cmd.CommandText = "truncate transacao_part";
+            await cmd.ExecuteNonQueryAsync();
+        }
+
         AnsiConsole.MarkupLine("[gray]OK[/]");
         
         AnsiConsole.Markup("[gray]Executar vacuum...[/]");
